bpf, io_uring: Fix compilation errors with -Wdefault-const-init-var-unsafe and incompatible-pointer-types
This commit addresses several compilation errors observed in `bpf_iter.c` and `io_uring.c` when compiling with stricter warnings.

The `default initialization of an object of type 'typeof ((attr->link_create.iter_info))' ... leaves the object uninitialized` warning, stemming from the `u64_to_user_ptr` macro and the `typecheck` helper, is resolved by explicitly casting the arguments to `(u64)`. This ensures that the type is properly inferred and avoids the uninitialized variable warning.

Additionally, in `io_uring.c`, the `incompatible pointer types passing 'struct timespec64 *' to parameter of type 'const struct __kernel_timespec *'` error is fixed. This occurs because `get_timespec64` expects a `const struct __kernel_timespec __user *` but was receiving `struct timespec64 __user *`. The solution involves casting the result of `u64_to_user_ptr` to the correct `(const struct __kernel_timespec __user *)` type, ensuring type compatibility with the function signature.

These changes ensure the kernel modules compile without the reported errors and warnings.

+++ kernel/bpf/bpf_iter.c
@@ -475,7 +475,7 @@ int bpf_iter_link_attach(const union bpf_attr *attr, struct bpf_prog *prog)
 
 	memset(&linfo, 0, sizeof(union bpf_iter_link_info));
 
-	ulinfo = u64_to_user_ptr(attr->link_create.iter_info);
+	ulinfo = u64_to_user_ptr((u64)attr->link_create.iter_info);
 	linfo_len = attr->link_create.iter_info_len;
 	if (!ulinfo ^ !linfo_len)
 		return -EINVAL;
